+++ examples/VPC_single_instance_in_subnet.py
@@ -44,7 +44,7 @@ instances',
         MinLength='9',
         MaxLength='18',
         Default='0.0.0.0/0',
-        AllowedPattern="(\d{1,3})\.(\d{1,3})\.(\d{1,3})\.(\d{1,3})/(\d{1,2})",
+        AllowedPattern=r"(\d{1,3})\.(\d{1,3})\.(\d{1,3})\.(\d{1,3})/(\d{1,2})",
         ConstraintDescription=(
             "must be a valid IP CIDR range of the form x.x.x.x/x."),
     ))